chore: BodyShop 락 전용 테스트 클래스 작성 #155
